# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AirportIQ API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/airports/{code}": {
      "get": {
        "summary": "Get Airport Metrics",
        "operationId": "get_airport_metrics_api_airports__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AirportMetricsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/airports/{code}/score": {
      "get": {
        "summary": "Get Airport Score",
        "operationId": "get_airport_score_api_airports__code__score_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AirportScore"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/airports/{code}/long-haul": {
      "get": {
        "summary": "Get Airport Long Haul",
        "operationId": "get_airport_long_haul_api_airports__code__long_haul_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "threshold_miles",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 3000.0,
              "title": "Threshold Miles"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Airport Long Haul Api Airports  Code  Long Haul Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/airports/compare": {
      "post": {
        "summary": "Compare Airports",
        "operationId": "compare_airports_api_airports_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/regions/{region}/ranking": {
      "get": {
        "summary": "Get Region Ranking",
        "operationId": "get_region_ranking_api_regions__region__ranking_get",
        "parameters": [
          {
            "name": "region",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Region"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Absent": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "absent",
            "title": "Kind",
            "default": "absent"
          },
          "reason": {
            "type": "string",
            "enum": [
              "NOT_PUBLISHED",
              "PROVIDER_FAILED",
              "NO_DEPARTURES",
              "OUT_OF_SCOPE"
            ],
            "title": "Reason"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "attempted": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Attempted"
          }
        },
        "type": "object",
        "required": [
          "reason",
          "detail",
          "attempted"
        ],
        "title": "Absent"
      },
      "AirportContext": {
        "properties": {
          "nas_delay_program": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DelayProgram"
              },
              {
                "type": "null"
              }
            ]
          },
          "weather": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WeatherContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "facility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FacilityContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "national_traffic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NationalTrafficContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "long_haul_departures": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "live",
                  "unavailable"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Long Haul Departures"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          }
        },
        "type": "object",
        "title": "AirportContext"
      },
...
      "AirportMetrics": {
        "properties": {
          "airport_code": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "pattern": "^[A-Z]{3}$",
            "title": "Airport Code"
          },
          "airport_name": {
            "type": "string",
            "title": "Airport Name",
            "default": ""
          },
          "passenger_volume": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Present_int_"
              },
              {
                "$ref": "#/components/schemas/Absent"
              }
            ],
            "title": "Passenger Volume"
          },
          "previous_passenger_volume": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Present_int_"
              },
              {
                "$ref": "#/components/schemas/Absent"
              }
            ],
            "title": "Previous Passenger Volume"
          },
          "annual_operations": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Present_int_"
              },
              {
                "$ref": "#/components/schemas/Absent"
              }
            ],
            "title": "Annual Operations"
          },
          "delayed_flights_pct": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Present_float_"
              },
              {
                "$ref": "#/components/schemas/Absent"
              }
            ],
            "title": "Delayed Flights Pct"
          },
          "average_delay_minutes": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Present_float_"
              },
              {
                "$ref": "#/components/schemas/Absent"
              }
            ],
            "title": "Average Delay Minutes"
          },
          "long_haul_flights": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Present_int_"
              },
              {
                "$ref": "#/components/schemas/Absent"
              }
            ],
            "title": "Long Haul Flights"
          },
          "total_departures": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Present_int_"
              },
              {
                "$ref": "#/components/schemas/Absent"
              }
            ],
            "title": "Total Departures"
          },
          "international_departures": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Present_int_"
              },
              {
                "$ref": "#/components/schemas/Absent"
              }
            ],
            "title": "International Departures",
            "default": {
              "kind": "absent",
              "reason": "NOT_PUBLISHED",
              "detail": "default",
              "attempted": []
            }
          },
          "average_flight_distance_sm": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Present_float_"
              },
              {
                "$ref": "#/components/schemas/Absent"
              }
            ],
            "title": "Average Flight Distance Sm",
            "default": {
              "kind": "absent",
              "reason": "NOT_PUBLISHED",
              "detail": "default",
              "attempted": []
            }
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources",
            "readOnly": true
          },
          "data_period": {
            "type": "string",
            "title": "Data Period",
            "readOnly": true
          },
          "core_completeness_score": {
            "type": "number",
            "title": "Core Completeness Score",
            "readOnly": true
          },
          "completeness_score": {
            "type": "number",
            "title": "Completeness Score",
            "readOnly": true
          },
          "freshness": {
            "type": "string",
            "enum": [
              "live",
              "sample",
              "mixed"
            ],
            "title": "Freshness",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "airport_code",
          "passenger_volume",
          "previous_passenger_volume",
          "annual_operations",
          "delayed_flights_pct",
          "average_delay_minutes",
          "long_haul_flights",
          "total_departures",
          "sources",
          "data_period",
          "core_completeness_score",
          "completeness_score",
          "freshness"
        ],
        "title": "AirportMetrics"
      },
      "AirportMetricsResponse": {
        "properties": {
          "metrics": {
            "$ref": "#/components/schemas/AirportMetrics"
          },
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AirportContext"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "metrics"
        ],
        "title": "AirportMetricsResponse"
      },
      "AirportScore": {
        "properties": {
          "airport_code": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "pattern": "^[A-Z]{3}$",
            "title": "Airport Code"
          },
          "airport_name": {
            "type": "string",
            "title": "Airport Name",
            "default": ""
          },
          "opportunity_score": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Present_float_"
              },
              {
                "$ref": "#/components/schemas/Absent"
              }
            ],
            "title": "Opportunity Score"
          },
          "demand_growth_score": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Present_float_"
              },
              {
                "$ref": "#/components/schemas/Absent"
              }
            ],
            "title": "Demand Growth Score"
          },
          "congestion_score": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Present_float_"
              },
              {
                "$ref": "#/components/schemas/Absent"
              }
            ],
            "title": "Congestion Score"
          },
          "delay_pressure_score": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Present_float_"
              },
              {
                "$ref": "#/components/schemas/Absent"
              }
            ],
            "title": "Delay Pressure Score"
          },
          "capacity_pressure_score": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Present_float_"
              },
              {
                "$ref": "#/components/schemas/Absent"
              }
            ],
            "title": "Capacity Pressure Score"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW"
            ],
            "title": "Confidence"
          },
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AirportContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "snapshot_at": {
            "type": "string",
            "title": "Snapshot At"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions",
            "readOnly": true
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations",
            "readOnly": true
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "airport_code",
          "opportunity_score",
          "demand_growth_score",
          "congestion_score",
          "delay_pressure_score",
          "capacity_pressure_score",
          "confidence",
          "snapshot_at",
          "assumptions",
          "limitations",
          "sources"
        ],
        "title": "AirportScore"
      },
...
      "CompareRequest": {
        "properties": {
          "airport_codes": {
            "items": {
              "type": "string",
              "maxLength": 3,
              "minLength": 3,
              "pattern": "^[A-Z]{3}$"
            },
            "type": "array",
            "title": "Airport Codes"
          }
        },
        "type": "object",
        "required": [
          "airport_codes"
        ],
        "title": "CompareRequest"
      },
      "ComparisonResult": {
        "properties": {
          "airports": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Airports"
          },
          "snapshot_at": {
            "type": "string",
            "title": "Snapshot At"
          },
          "kind": {
            "type": "string",
            "const": "comparison",
            "title": "Kind",
            "default": "comparison"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ComparisonRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "highest_congestion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highest Congestion"
          },
          "highest_opportunity_score": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highest Opportunity Score"
          },
          "period_warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Warning"
          }
        },
        "type": "object",
        "required": [
          "airports",
          "snapshot_at",
          "rows",
          "highest_congestion",
          "highest_opportunity_score",
          "period_warning"
        ],
        "title": "ComparisonResult"
      },
      "ComparisonRow": {
        "properties": {
          "airport_code": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "pattern": "^[A-Z]{3}$",
            "title": "Airport Code"
          },
          "score": {
            "$ref": "#/components/schemas/AirportScore"
          },
          "metrics": {
            "$ref": "#/components/schemas/AirportMetrics"
          },
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AirportContext"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "airport_code",
          "score",
          "metrics"
        ],
        "title": "ComparisonRow"
      },
...
      "DelayProgram": {
        "properties": {
          "airport_icao": {
            "type": "string",
            "title": "Airport Icao"
          },
          "program": {
            "type": "string",
            "title": "Program"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "avg_delay": {
            "type": "string",
            "title": "Avg Delay"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "airport_icao",
          "program",
          "reason",
          "avg_delay",
          "updated_at"
        ],
        "title": "DelayProgram"
      },
...
      "FacilityContext": {
        "properties": {
          "acreage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acreage"
          },
          "part139_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part139 Class"
          },
          "tower_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tower Type"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "effective_date": {
            "type": "string",
            "title": "Effective Date"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "NTAD Aviation Facilities"
          }
        },
        "type": "object",
        "required": [
          "acreage",
          "part139_class",
          "tower_type",
          "status",
          "effective_date"
        ],
        "title": "FacilityContext"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Live": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "live",
            "title": "Kind",
            "default": "live"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "fetched_at": {
            "type": "string",
            "title": "Fetched At"
          }
        },
        "type": "object",
        "required": [
          "source",
          "period",
          "fetched_at"
        ],
        "title": "Live"
      },
...
      "Present_float_": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "present",
            "title": "Kind",
            "default": "present"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "origin": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Live"
              },
              {
                "$ref": "#/components/schemas/Sample"
              },
              {
                "$ref": "#/components/schemas/Derived"
              },
              {
                "$ref": "#/components/schemas/Proxy"
              },
              {
                "$ref": "#/components/schemas/Scenario"
              }
            ],
            "title": "Origin"
          }
        },
        "type": "object",
        "required": [
          "value",
          "origin"
        ],
        "title": "Present[float]"
      },
      "Present_int_": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "present",
            "title": "Kind",
            "default": "present"
          },
          "value": {
            "type": "integer",
            "title": "Value"
          },
          "origin": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Live"
              },
              {
                "$ref": "#/components/schemas/Sample"
              },
              {
                "$ref": "#/components/schemas/Derived"
              },
              {
                "$ref": "#/components/schemas/Proxy"
              },
              {
                "$ref": "#/components/schemas/Scenario"
              }
            ],
            "title": "Origin"
          }
        },
        "type": "object",
        "required": [
          "value",
          "origin"
        ],
        "title": "Present[int]"
      },
...
      "RankingResult": {
        "properties": {
          "airports": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Airports"
          },
          "snapshot_at": {
            "type": "string",
            "title": "Snapshot At"
          },
          "kind": {
            "type": "string",
            "const": "ranking",
            "title": "Kind",
            "default": "ranking"
          },
          "ranked": {
            "items": {
              "$ref": "#/components/schemas/AirportScore"
            },
            "type": "array",
            "title": "Ranked"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "peer_note": {
            "type": "string",
            "title": "Peer Note"
          },
          "metric": {
            "type": "string",
            "title": "Metric",
            "default": "opportunity_score"
          }
        },
        "type": "object",
        "required": [
          "airports",
          "snapshot_at",
          "ranked",
          "region",
          "peer_note"
        ],
        "title": "RankingResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeatherContext": {
        "properties": {
          "flight_category": {
            "type": "string",
            "title": "Flight Category"
          },
          "wind": {
            "type": "string",
            "title": "Wind"
          },
          "visibility": {
            "type": "string",
            "title": "Visibility"
          },
          "remarks": {
            "type": "string",
            "title": "Remarks"
          },
          "observation_time": {
            "type": "string",
            "title": "Observation Time"
          }
        },
        "type": "object",
        "required": [
          "flight_category",
          "wind",
          "visibility",
          "remarks",
          "observation_time"
        ],
        "title": "WeatherContext"
      },
...
    }
  }
}